refactor(viewer): address remaining PR review minors in SessionRelationshipsView
- Add SessionScanWireData to @vibe-replay/types as the canonical
  /api/scan/results wire type; useRelationshipData re-exports it as
  ScanResultSession so the viewer hook no longer has its own copy
  that can drift from the scanner's SessionScanResult definition.
- Extract shared `timeAgo` and `shortName` into utils/format.ts
  (timeAgo gains a "long" / "short" format flag); ProjectsPanel and
  SessionRelationshipsView now import from there instead of each
  defining their own near-duplicates.
- Replace the magic 300000 dispatch grace window with a named
  DISPATCH_END_GRACE_MS constant and a comment explaining why a
  child session may legitimately overrun the parent's recorded end.
