+++ packages/viewer/src/components/ProjectsPanel.tsx
@@ -4,6 +4,7 @@
 
 import { useMemo, useState } from "react";
 import { localDayKey } from "../utils/date";
+import { timeAgo } from "../utils/format";
 import { navigateTo, rollupTopProjects } from "./dashboard-utils";
 import { useScanInsightsContext } from "./InsightsPanel";
 import SessionRelationshipsView from "./SessionRelationshipsView";
@@ -56,20 +57,6 @@ function fmtNum(n: number): string {
   if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
   if (n >= 1_000) return `${(n / 1_000).toFixed(1)}K`;
   return String(n);
-}
-
-function timeAgo(iso: string): string {
-  if (!iso) return "";
-  const ms = Date.now() - new Date(iso).getTime();
-  const mins = Math.floor(ms / 60000);
-  if (mins < 1) return "just now";
-  if (mins < 60) return `${mins}m ago`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  if (days < 30) return `${days}d ago`;
-  const months = Math.floor(days / 30);
-  return `${months}mo ago`;
 }
 
 // ─── Main Component ─────────────────────────────────────────────────
@@ -216,7 +203,7 @@ export default function ProjectsPanel({ onNavigate }: ProjectsPanelProps) {
                   </div>
                   {p.lastActivity && (
                     <span className="text-[10px] font-mono text-terminal-dimmer shrink-0">
-                      {timeAgo(p.lastActivity)}
+                      {timeAgo(p.lastActivity, "long")}
                     </span>
                   )}
                 </div>
